Do not convert to `int` after flatfield correction

+++ instamatic/processing/flatfield.py
@@ -52,13 +52,12 @@ def apply_flatfield_correction(img, flatfield, darkfield=None):
     Apply flatfield correction to image
 
     https://en.wikipedia.org/wiki/Flat-field_correction"""
-    
     if darkfield is None:
         ret = img * np.mean(flatfield) / flatfield
     else:
         gain = np.mean(flatfield - darkfield) / (flatfield - darkfield)
         ret = (img - darkfield) * gain
-    return ret.astype(int)
+    return ret
 
 
 def collect_flatfield(ctrl=None, frames=100, save_images=False, **kwargs):
